move sprinkler logging into storage
