+++ main.cpp
@@ -475,8 +475,6 @@ static bool process_special_program_command(const char*, uint32_t curr_time);
 static void perform_ntp_sync();
 
 #if defined(ESP8266)
-bool delete_log_oldest();
-uint32_t get_sprinkler_log_size();
 void start_server_ap();
 void start_server_client();
 static Ticker reboot_ticker;
@@ -1135,223 +1133,6 @@ void check_weather() {
 		#endif
 		GetWeather();
 	}
-}
-
-	// ================================
-// ====== LOGGING FUNCTIONS =======
-// ================================
-
-/** Generate log file name
- * Log files will be named /logs/xxxxx.txt
- */
-void make_logfile_name(char *name) {
-	strcpy(tmp_buffer+TMP_BUFFER_SIZE-10, name); // hack: we do this because name is from tmp_buffer too
-	strcpy(tmp_buffer, LOG_DIR);
-	strcat(tmp_buffer, tmp_buffer+TMP_BUFFER_SIZE-10);
-	strcat_P(tmp_buffer, PSTR(".txt"));
-}
-
-/* To save RAM space, we store log type names
- * in program memory, and each name
- * must be strictly two characters with an ending 0
- * so each name is 3 characters total
- */
-static const char log_type_names[] PROGMEM =
-	"  \0"
-	"s1\0"
-	"rd\0"
-	"wl\0"
-	"fl\0"
-	"s2\0"
-	"s3\0"
-	"s4\0"
-	"cu\0";
-
-/** write run record to log on SD card */
-void write_log(unsigned char type, time_os_t curr_time) {
-
-	if (!os.iopts[IOPT_ENABLE_LOGGING]) return;
-
-	// file name will be logs/xxxxx.tx where xxxxx is the day in epoch time
-	snprintf (tmp_buffer, TMP_BUFFER_SIZE, "%" PRIu32, (uint32_t)curr_time / 86400);
-	make_logfile_name(tmp_buffer);
-
-	// Step 1: open file if exists, or create new otherwise,
-	// and move file pointer to the end
-#if defined(ESP8266) // prepare log folder for Arduino
-	File file = LittleFS.open(tmp_buffer, "r+");
-	if(!file) {
-		// New day file — trim sprinkler log budget before creating it
-		uint32_t limit = (uint32_t)LOG_SPRINKLER_MAX_KB * 1024;
-		while (get_sprinkler_log_size() >= limit) {
-			if (!delete_log_oldest()) break;
-		}
-		file = LittleFS.open(tmp_buffer, "w");
-		if(!file) return;
-	}
-	file.seek(0, SeekEnd);
-#else // prepare log folder for RPI/LINUX
-	struct stat st;
-	if(stat(get_filename_fullpath(LOG_DIR), &st)) {
-		if(mkdir(get_filename_fullpath(LOG_DIR), S_IRUSR | S_IWUSR | S_IXUSR | S_IRGRP | S_IWGRP | S_IXGRP | S_IROTH | S_IWOTH | S_IXOTH)) {
-			return;
-		}
-	}
-	FILE *file;
-	file = fopen(get_filename_fullpath(tmp_buffer), "rb+");
-	if(!file) {
-		file = fopen(get_filename_fullpath(tmp_buffer), "wb");
-		if (!file)	return;
-	}
-	fseek(file, 0, SEEK_END);
-#endif	// prepare log folder
-
-	// Step 2: prepare data buffer
-	strcpy_P(tmp_buffer, PSTR("["));
-
-	if(type == LOGDATA_STATION) {
-		size_t size = strlen(tmp_buffer);
-		snprintf(tmp_buffer + size, TMP_BUFFER_SIZE - size , "%d", pd.lastrun.program);
-		strcat_P(tmp_buffer, PSTR(","));
-		size = strlen(tmp_buffer);
-		snprintf(tmp_buffer + size, TMP_BUFFER_SIZE - size , "%d", pd.lastrun.station);
-		strcat_P(tmp_buffer, PSTR(","));
-		// duration is unsigned integer
-		size = strlen(tmp_buffer);
-		snprintf(tmp_buffer + size, TMP_BUFFER_SIZE - size , "%" PRIu32, (uint32_t)pd.lastrun.duration);
-
-	} else {
-		uint32_t lvalue=0;
-		if(type==LOGDATA_FLOWSENSE) {
-			lvalue = (flow_count>os.flowcount_log_start)?(flow_count-os.flowcount_log_start):0;
-		}
-
-		size_t size = strlen(tmp_buffer);
-		snprintf(tmp_buffer + size, TMP_BUFFER_SIZE - size , "%" PRIu32, lvalue);
-		strcat_P(tmp_buffer, PSTR(",\""));
-		strcat_P(tmp_buffer, log_type_names+type*3);
-		strcat_P(tmp_buffer, PSTR("\","));
-
-		switch(type) {
-			case LOGDATA_FLOWSENSE:
-			case LOGDATA_SENSOR1:
-			case LOGDATA_SENSOR2:
-			case LOGDATA_SENSOR3:
-			case LOGDATA_SENSOR4: {
-				// FLOWSENSE and SENSOR1..4 all use the SN(i)_active_lasttime
-				// for their duration value; pick the right sensor index.
-				int8_t sidx = sensor_index_from_log_code(type);
-				time_os_t t = (sidx >= 0) ? os.sn_sensors[sidx].active_lasttime : 0;
-				lvalue = (curr_time>t) ? (curr_time-t) : 0;
-				break;
-			}
-			case LOGDATA_RAINDELAY:
-				lvalue = (curr_time>os.raindelay_on_lasttime)?(curr_time-os.raindelay_on_lasttime):0;
-				break;
-			case LOGDATA_WATERLEVEL:
-				lvalue = os.iopts[IOPT_WATER_PERCENTAGE];
-				break;
-		}
-		size = strlen(tmp_buffer);
-		snprintf(tmp_buffer + size, TMP_BUFFER_SIZE - size , "%" PRIu32, lvalue);
-	}
-	strcat_P(tmp_buffer, PSTR(","));
-	size_t size = strlen(tmp_buffer);
-	snprintf(tmp_buffer + size, TMP_BUFFER_SIZE - size , "%" PRIu32, (uint32_t)curr_time);
-	if((os.iopts[IOPT_SENSOR1_TYPE]==SENSOR_TYPE_FLOW) && (type==LOGDATA_STATION)) {
-		// RAH implementation of flow sensor
-		strcat_P(tmp_buffer, PSTR(","));
-		#if defined(ESP8266)
-		dtostrf(flow_last_gpm,5,2,tmp_buffer+strlen(tmp_buffer));
-		#else
-		snprintf(tmp_buffer+strlen(tmp_buffer), TMP_BUFFER_SIZE, "%5.2f", flow_last_gpm);
-		#endif
-	}
-	strcat_P(tmp_buffer, PSTR("]\r\n"));
-
-#if defined(ESP8266)
-	file.write((const uint8_t*)tmp_buffer, strlen(tmp_buffer));
-	file.close();
-#else
-	fwrite(tmp_buffer, 1, strlen(tmp_buffer), file);
-	fclose(file);
-#endif
-}
-
-#if defined(ESP8266)
-uint32_t get_sprinkler_log_size() {
-	Dir dir = LittleFS.openDir(LOG_DIR);
-	uint32_t total = 0;
-	while (dir.next()) {
-		if (dir.fileName().endsWith(".txt"))
-			total += dir.fileSize();
-	}
-	return total;
-}
-
-bool delete_log_oldest() {
-	Dir dir = LittleFS.openDir(LOG_DIR);
-	uint32_t oldest_day = UINT32_MAX;
-	String oldest_fn;
-	while (dir.next()) {
-		if (!dir.fileName().endsWith(".txt")) continue;
-		uint32_t day = (uint32_t)atol(dir.fileName().c_str());
-		if (day < oldest_day) {
-			oldest_day = day;
-			oldest_fn = dir.fileName();
-		}
-	}
-	if (oldest_fn.length() > 0) {
-		DEBUG_PRINT(F("deleting "))
-		DEBUG_PRINTLN(LOG_DIR + oldest_fn);
-		LittleFS.remove(LOG_DIR + oldest_fn);
-		return true;
-	}
-	return false;
-}
-#endif
-
-/** Delete log file
- * If name is 'all', delete all logs
- */
-void delete_log(char *name) {
-	if (!os.iopts[IOPT_ENABLE_LOGGING]) return;
-#if defined(ESP8266)
-	if (strncmp(name, "all", 3) == 0) {
-		// delete all sprinkler log files (.txt), leaving sensor logs intact
-		Dir dir = LittleFS.openDir(LOG_DIR);
-		while (dir.next()) {
-			if (dir.fileName().endsWith(".txt"))
-				LittleFS.remove(LOG_DIR+dir.fileName());
-		}
-	} else {
-		// delete a single log file
-		make_logfile_name(name);
-		if(!LittleFS.exists(tmp_buffer)) return;
-		LittleFS.remove(tmp_buffer);
-	}
-#else // delete_log implementation for RPI/LINUX
-	if (strncmp(name, "all", 3) == 0) {
-		// delete all sprinkler log files (.txt), leaving sensor logs intact
-		char log_dir[PATH_MAX];
-		strcpy(log_dir, get_filename_fullpath(LOG_DIR));
-		DIR *d = opendir(log_dir);
-		if (d) {
-			int log_dir_fd = dirfd(d);
-			struct dirent *ent;
-			while ((ent = readdir(d)) != NULL) {
-				size_t len = strlen(ent->d_name);
-				if (len > 4 && strcmp(ent->d_name + len - 4, ".txt") == 0) {
-					unlinkat(log_dir_fd, ent->d_name, 0);
-				}
-			}
-			closedir(d);
-		}
-	} else {
-		make_logfile_name(name);
-		remove(get_filename_fullpath(tmp_buffer));
-	}
-#endif
 }
 
 /** Perform network check
